Install Console_Table package via pear
Drush requires Console_Table to run. Download it by default so we
do not have to run drush with sudo just to download the package.
Otherwise, we may end up with the following error:

vagrant@lucid32:~$ drush status
The drush command 'status' could not be found.                       [error]
Drush needs a copy of the PEAR Console_Table library in order to     [error]
function, and the attempt to download this file automatically failed
because you do not have permission to write files in
/usr/share/php/drush/includes. To continue you will need to download
the 1.1.3 package from http://pear.php.net/package/Console_Table,
extract it, and copy the Table.php file into Drush's directory as
/usr/share/php/drush/includes/table.inc.