Link /home/vagrant/.drush to /vagrant/.drush
If you are running vagrant it is vital to have access to drush aliases.  In
order to make them part of the vagrant environment we link them in from the
project.  That way you can share them.  Furture concept would be to generate
the site aliase automatically for the @dev environment.

It only performs the link if the .drush directory exists.