Remove subversion from suggested cookbook list.
Drush provides integration with Subversion; but also with Bazaar, curl
or wget, rsync, gzip, open or xdg-open, etc, and presumably other
programs in the future.

IMO, we should either include suggestions for the aforementioned
programs or avoid referencing programs that work with Drush altogether.

If we avoid referencing programs that work with Drush, then we can avoid
chasing Drush HEAD as often.